refactor: changed pdf-viewer to global component
