+++ packages/front-end/app/view/[sessionId]/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 import "./_util/pdfWorkerPolyfill";
-import PDFViewer from "./_components/PDFViewer";
+import PDFViewer from "@/components/DocumentViewer/PDF/PDFViewer";
 import {useQuery} from "@tanstack/react-query";
 import { AxiosResponse } from "axios";
 import { Session, File } from "@/schema/backend.schema";
